move FeedStoreSpecs protocols to a separate file
