+++ EssentialFeedTests/Feed Cache/CodableFeedStoreTests.swift
@@ -1,37 +1,7 @@
 import XCTest
 import EssentialFeed
 
-protocol FeedStoreSpecs {
-    func test_retrieve_deliversEmptyOnNonEmptyCache()
-    func test_retrieve_hasNoSideEffectsOnEmptyCache()
-    func test_retrieve_deliversDataOnNonEmptyCache()
-    func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
-    func test_insert_deliversNoErrorOnEmptyCache()
-    func test_insert_deliversNoErrorOnNonEmptyCache()
-    func test_insert_overridesPreviouslyInsertedValues()
-    func test_delete_deliversNoErrorOnEmptyCache()
-    func test_delete_deliversNoErrorOnNonEmptyCache()
-    func test_delete_hasNoSideEffectsOnEmptyCache()
-    func test_delete_emptiesPreviouslyInsertedCache()
-    func test_storeSideEffects_runSerially()
-}
-
-protocol FailableRetrieveStoreSpecs {
-    func test_retrieve_deliversErrorOnRetrievalFailure()
-    func test_retrieve_hasNoSideEffectsOnFailure()
-}
-
-protocol FailableInsertStoreSpecs {
-    func test_insert_deliversErrorOnInsertionFailure()
-    func test_insert_hasNoSideEffectsOnFailure()
-}
-
-protocol FailableDeleteStoreSpecs {
-    func test_delete_deliversErrorOnDeletionFailure()
-    func test_delete_hasNoSideEffectsOnFailure()
-}
-
-class CodableFeedStoreTests: XCTestCase, FailableRetrieveStoreSpecs, FailableInsertStoreSpecs, FailableDeleteStoreSpecs {
+class CodableFeedStoreTests: XCTestCase, FailableFeedStoreSpec {
     override func setUp() {
         super.setUp()
         clearTestStore()
